Расскидывает и «предугадывает» модули
