+++ js/choose.js
@@ -1,0 +1,1 @@
+// for choosing offer on a map
