+++ js/comparison.js
@@ -1,0 +1,1 @@
+// for offer comparison
