+++ js/form-submit.js
@@ -1,0 +1,1 @@
+// for form submiting anf getting offers
