+++ js/page-state.js
@@ -1,0 +1,1 @@
+// for setting page state - active||none
